Move stylesheets to `/styles` directory
This commit moves all the CSS styles to `/styles` directory and adds a
couple of gulp tasks to generate these styles on prepublish. This
results in having a `/dist` directory being used just for distributing
the package and development. Currently the task has to be run manually
when developing.

Fixes mapillary/mapillary-js#2