Indicate if mesh is populated.
Empty meshes can be returned for merged nodes. They will parse
to null. To be able to know that a mesh has been loaded but
was empty the populated bool is added to the mesh interface
and set in the mesh loading. If the mesh is null it falls
back to a non populated mesh with empty vertex and face
arrays.