Show max image size texture on halt.
Delay high res texture creation for 2000 milliseconds
to avoid setting to fast when navigating.
Abort if a new node item is emitted in stream by
flat mapping latest which disposes the previous
observable.
Dispose old texture before setting the new one.
Ensure that node is the same as the currently displayed.