Retreive key instead of node for next node.
When a node is outside all retrieved tiles the result when
requesting it from the graph is null. This can happen for
sequence edges because they are not spatially related.
When retrieving the key for a sequence edge with a node
that does not exist in the graph a tiles fetch will be
triggered that will result in the node being added to
the graph.