Render when single hash is added or cleared.
When the first hash is added to the GL renderer the
frame subscription is initiated. It will not render
for the frame of the render hash but needs to render
for the next frame. Therefor needs render is set on
the render camera.
Clear hash before clearing renderer to ensure that
one empty render is drawn.
Unit test when clearing one of multiple renders.