Initialize component controller correctly when no key provided.
Fixes #130
Fixes #143
Fixes #144

Always initialize components in controller constructor to ensure
that their state is not overwritten later if changed by API calls.
Do not show cover component when viewer is not initialized with a
key.
Start state when a node has been set on the state.
Do not use cover component to set the node a second time after
initialization.