Consistent handling of modifier keys.
If alt or shift is pressed the navigation is now consistent
across edge directions. Step directions are used when no
modifier key is pressed, turn directions when shift is pressed
and sequense directions when alt is pressed.
If shift or alt is pressed and there is no node in the correct
direction there is no fallback to step directions.