Ensure that fetch/fill errors are propagated.
Possible errors from fetch and fill must be propagated to
the caller to be able to catch them. When the calls are
made outside of the stream and the graph is scanned with
skipWhile the errors are left unhandled and not propagated
to the caller.

Ensure that multiple consecutive calls work on the same stream
to avoid running graph logic multiple times.

Adjust unit tests to new logic.