feat(popup): handle selection in all browsers
Ensure that selection works in Safari, Chrome and Firefox.
Content like text in popups is selectable, virtual dom
rendered elements are not selectable.
Prevent selection when panning as well as when dragging from
a dom element.
Subscribe to events on dom container besides canvas container
to ensure that events does not affect popup container where
selection is enabled.
Ensure that default is prevented for all viewer elements
except popups for touch events so that dragging does not
cause page scroll etc.