fix: combine spatial edges and sequence observable
Fixes #257

Use combine latest instead of with latest from because
with latest from does not emit at all if the target observable
does not have a stored last emitted item. This is sometimes the
case when the node navigated to is not cached and therefore its
sequence is not cashed. Only a problem when distinguishSequence
option is enabled.