fix: support kayboard navigation in firefox
Set tab index attribute on canvas instead of container to
ensure that focues is set in Firefox. Firefox does not
apply focus to an element that has zero size currently.
This results in the fact that keyboard events will not
be emitted when using fallback functionality by not
enabling WebGL components so that the WebGL renderer is
never created and canvas container has no size.