refactor: avoid creating an unused interactive canvas
When no WebGL components are activated the canvas should either
not be created at all  or created and maintained. If the canvas
is not resized correctly it will behave incorrectly. Therefore
do not create it at all for non WebGL state. This will result in
that interactive components like mouse and keyboard will not
function for non WebGL functionality.