feat: cache trajectory nodes on graph mode change
Cache current node if graph mode changes to sequence
and all upcoming trajectory nodes in parallel if
graph mode changes to spatial to retrieve the
spatial edges for the upcoming nodes.
Set timeouts on edges to ensure that observables
does not get stuck on missing spatial edges if
tile caching fails.