fix: do not pre-cache same node multiple times
Fixes #270

Ensure that the first node cached is the node directly after the
previously last cached node.
Ensure that the last request key index is not outside of array
which lead to last request key becoming undefined and the same
nodes to be pre-cached for every state frame until successfully
appended on state.
Test that sequence is reversed when playing in prev direction.