feat: bridge sequence play though camera uuid
If nodes with the same camera uuid exists in the vicinity of
the last node in the sequence with roughly the same timestamp
play is bridged from the last node in the sequence to the best
match with respect to delta time.
Update unit tests to account for new logic.