feat: handle 2d and 3d domains for tags
Polygons can have straight lines between vertices either
in the distorted 2D domain of the image or in the
undistorted 3D domain of the viewer space.
Add option to outline tag for selecting domain. Default
to 2D domain to keep existing behavior compatible.
Polygons in 2D domain are subsampled and can therefore not
be editable.